# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orange Banking API",
        "description": "API documentation for Orange Banking System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Get authenticated client's accounts",
                "description": "Get all accounts belonging to the authenticated client",
                "operationId": "getMyAccounts",
                "responses": {
                    "200": {
                        "description": "Accounts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comptes récupérés"
                                        },
                                        "data": {
                                            "properties": {
                                                "client": {
                                                    "$ref": "#/components/schemas/Client"
                                                },
                                                "comptes": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Compte"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/comptes/{id}/solde": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Get balance for specific account",
                "description": "Get balance for a specific account by ID (user must own the account)",
                "operationId": "getAccountBalanceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balance retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solde récupéré avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "compte_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "numero_compte": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès non autorisé à ce compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte introuvable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/comptes/{id}/historique": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Get transaction history for specific account",
                "description": "Get all transactions for a specific account by ID (user must own the account)",
                "operationId": "getAccountTransactionHistoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction history retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Historique des transactions récupéré"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès non autorisé à ce compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte introuvable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/comptes/solde": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Get account balances",
                "description": "Get balances for all accounts of the authenticated user",
                "operationId": "getAccountBalances",
                "responses": {
                    "200": {
                        "description": "Balances retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Soldes récupérés avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "compte_id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "numero_compte": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "solde": {
                                                        "type": "number",
                                                        "format": "float"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "f9136320-10c8-47cc-8de0-ba4cc2ef78b1"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Dupont"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Jean"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jean.dupont@example.com"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "0123456789"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Compte": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "f9136320-10c8-47cc-8de0-ba4cc2ef78b1"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "numero_compte": {
                        "type": "string",
                        "example": "FR4284166632"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": "150.75"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "courant",
                            "cheque",
                            "epargne"
                        ],
                        "example": "courant"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}